add hystogram function with file
